add  UART and ENCODER

+++ LowLevel/Drivers/HAL/L289/L289_Int.h
@@ -9,7 +9,7 @@
 
 void HL289_vInit(GPIO_PinConfig_t* dir1,GPIO_PinConfig_t* dir2,GPIO_PinConfig_t* en);
 void HL289_vTIM_Init(Timer_Config_t* TimerCfg,Timer_OutputCompare_Config_t* OCcfg,TIMER_TypeDef *TIMERx ,u32 CHx);
-void HL289_vSetSpeed(GPIO_PinConfig_t* dir1,GPIO_PinConfig_t* dir2,Timer_Config_t* TimerCfg,u8 speed);
+void HL289_vSetSpeed(GPIO_PinConfig_t* dir1,GPIO_PinConfig_t* dir2,Timer_Config_t* TimerCfg,s16 speed);
 
 
 #endif
